requirements.txt
aioboto3==12.3.0
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
blinker==1.7.0
boto3==1.34.34
botocore==1.34.34
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai==1.12.0
openai-whisper==20231117
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
SQLAlchemy==2.0.27
sympy==1.12
tiktoken==0.6.0
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
